[day8]

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Post;
+use App\Comment;
 use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
@@ -12,9 +13,9 @@ class PostController extends Controller
     public function index()
     {
 
-        $posts=Post::orderby('created_at','desc')->paginate(6);
+        $posts = Post::orderby('created_at', 'desc')->withCount('comments')->paginate(6);
 
-        return view('post.index',compact('posts'));
+        return view('post.index', compact('posts'));
     }
 
     public function create()
@@ -27,41 +28,45 @@ class PostController extends Controller
     {
 
         //验证
-        $this->validate(request(),[
-            'title'=>'required|string|max:100|min:5',
-            'content'=>'required|string|min:10',
+        $this->validate(request(), [
+            'title' => 'required|string|max:100|min:5',
+            'content' => 'required|string|min:10',
         ]);
 
-        $user_id=Auth::id();
-        $params = array_merge(request(['content','title']),compact('user_id'));
+        $user_id = Auth::id();
+        $params = array_merge(request(['content', 'title']), compact('user_id'));
 
-        $post=Post::create($params);
+        $post = Post::create($params);
         return redirect('/posts');
 
     }
 
     public function show(Post $post)
     {
-        return view('post.show',compact('post'));
+        //预加载评论
+        $post->load('comments');
+        return view('post.show', compact('post'));
 
     }
 
     public function edit(Post $post)
     {
-        return view('post.edit',compact('post'));
+        return view('post.edit', compact('post'));
     }
 
     public function update(Post $post)
     {
         //验证
-        $this->validate(request(),[
-            'title'=>'required|string|max:100|min:5',
-            'content'=>'required|string|min:10',
+        $this->validate(request(), [
+            'title' => 'required|string|max:100|min:5',
+            'content' => 'required|string|min:10',
         ]);
+        //权限验证
+        $this->authorize('update', $post);
 
         //保存
-        $post->title=request('title');
-        $post->content=request('content');
+        $post->title = request('title');
+        $post->content = request('content');
         $post->save();
 
         return redirect("/posts/{$post->id}");
@@ -69,6 +74,8 @@ class PostController extends Controller
 
     public function delete(Post $post)
     {
+        //权限验证
+        $this->authorize('delete', $post);
         $post->delete();
 
         return redirect('/posts');
@@ -76,7 +83,24 @@ class PostController extends Controller
 
     public function imageUpload(Request $request)
     {
-        $path=$request->file('wangEditorH5File')->storePublicly(md5(time()));
-        return asset('storage/'.$path);
+        $path = $request->file('wangEditorH5File')->storePublicly(md5(time()));
+        return asset('storage/' . $path);
+    }
+
+    //评论
+    public function comment(Post $post)
+    {
+        $this->validate(request(), [
+            'content' => 'required|min:3',
+        ]);
+
+        $comment = new Comment();
+        $comment->user_id= Auth::id();
+        $comment->content= request('content');
+
+        $post->comments()->save($comment);
+
+        return back();
+
     }
 }
